ETL/requirements.txt
psycopg2-binary==2.9.5
backoff==2.2.1
redis==4.3.5
pydantic==1.9.0
python-dotenv==0.21.0
elasticsearch==7.9.1